fix review ui
